Offer more control over reqwest features
Currently, the reqwest dependency picks up all default features, enabling native-tls, which depends on OpenSSL. For projects that do not want to have an OpenSSL dependency, this is a problem.

This PR turns off default features for the reqwest dependency. Dependent projects can now choose whichever features of reqwest they want to enable (they need a dependency on reqwest to construct a `Validator`, anyway, so that is a natural place to customize its features).